PredictionServer/requirements_npu.txt
torch>=1.6
furiosa-sdk[litmus]
onnx
onnx-simplifier
numpy
pandas
fair-esm
